+++ java/android/app/src/main/java/com/github/zaolahma/robotinterface/core/comm/protocol/UnlockInd.java
@@ -1,7 +1,7 @@
 package com.github.zaolahma.robotinterface.core.comm.protocol;
 
 public class UnlockInd implements Message {
-    public static final int S_MESSAGE_ID = 10;
+    public static final byte S_MESSAGE_ID = 10;
 
     @Override
     public byte[] encode() {
@@ -16,7 +16,7 @@ public class UnlockInd implements Message {
     }
 
     @Override
-    public int getMessageId() {
+    public byte getMessageId() {
         return S_MESSAGE_ID;
     }
 }
